% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study.R
\name{study_accuracy}
\alias{study_accuracy}
\title{Select models based on validation performance and conduct an evaluation study}
\usage{
study_accuracy(load.folder = "G:/R/MLE_SIM/DATA", load.base = "job",
  load.id = 8, methods = NA, M = 200, M.start = NA,
  M.probs = "uniform", M.seed = 1, eval.first = 1, eval.n = 200,
  eval.rdm = FALSE, estimate.method = "beta.approx",
  estimate.args = "", select.method = "rank", select.args = "r=1",
  select.limit = "none", known.delta = FALSE, infer.method = "maxT",
  alternative = "greater", alpha = 0.025, transform = "none",
  data = NULL, instance = NULL, job = NULL)
}
\arguments{
\item{load.folder}{character, path containing data instances}

\item{load.base}{character, define names of data instances (default="job")}

\item{load.id}{integer, the data instance with name paste0(load.base, load.id, ".RData")) is loaded
from directory load.folder}

\item{methods}{character, potentially subset available prediction models by method (=learning algorithm)
e.g. recover elastic net models by specifying methods="glmnet" (caret train.method),
no effect if methods=NA (default)}

\item{M}{integer, number of models to subsample from available models (restricted via methods argument),
needs to be less or euqal than number of available models (200 per default)}

\item{M.start}{integer, starting index for subsetting}

\item{M.probs}{character, "uniform" for random subset, "learn.theta" for P(selected)=learn.theta(=true model performance),
"learn.theta.neg" for P(selected)=1-learn.theta}

\item{M.seed}{integer, seed for random subsetting (i.e. if M.probs != "uniform")}

\item{eval.first}{integer, index of first evaluation observation (from all available)}

\item{eval.n}{integer, test sample size}

\item{eval.rdm}{logical, choose test samples randomly? (default: FALSE)}

\item{estimate.method}{character, estimation method in SEPM package default ("beta.approx")}

\item{estimate.args}{character, specify additional estimation argument as character of form
"arg1=value1_arg2=value2_..."}

\item{select.method}{character, selection method based on validation ranking, e.g. "rank" (default) or "se"}

\item{select.args}{character, further arguments defining selection rule e.g. "r=1" for
select.method="rank" to choose only best validation models or "c=1" for select.method="se"
(which defines the 'within1SE# rule)}

\item{select.limit}{integer, maximum number of models to evaluate}

\item{known.delta}{logical, selection based on validation performance + known performance gain after re-training?
(default: FALSE)}

\item{infer.method}{character, defines the statistical test, e.g. "maxT", "Bonferroni" or "naive"}

\item{alternative}{charatcer, either "greater" (default), "lower" or "two.sided"}

\item{alpha}{numeric, significan level (default: 0.025)}

\item{transform}{character, specifies transformation of test statistics, passed to SEPM::evaluate()}

\item{data}{NULL, batchtools argument}

\item{job}{NULL, batchtools argument}
}
\value{
Returns a list which contains all relevant characteristics of the evaluation study.
}
\description{
This function reads in data instances produces via generate_data() and emulates the process of
conducting an evaluation study for one or multiple selected prediction models.
}
